requirements.txt
streamlit>=1.30.0
bcrypt>=4.0.0
pandas>=2.0.0
mysql-connector-python>=8.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
